structs in corect files
